PWM module changed to calculated resolution

+++ FreeRTOS/Demo/CORTEX_CY8C5588_PSoC_Creator_GCC/FreeRTOS_Demo.cydsn/currentposition.h
@@ -13,8 +13,8 @@
 #ifndef CURRENTPOSITION_H
 #define CURRENTPOSITION_H
     
-uint8_t usGetCurrentPosition( void );
-void    vSetCurrentPosition( uint8_t usNewPosition );
+uint16_t usGetCurrentPosition( void );
+void    vSetCurrentPosition( uint16_t usNewPosition );
     
 #endif
